requirements.txt
Babel==2.6.0
Django==2.1.3
django-phonenumber-field==2.1.0
mysql==0.0.2
mysqlclient==1.3.13
phonenumbers==8.9.16
pytz==2018.7
